requirements.txt
fitz==0.0.1.dev2
imbalanced_learn==0.13.0
imblearn==0.0
joblib==1.4.2
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.3
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
pytesseract==0.3.13
scikit_learn==1.6.1
seaborn==0.13.2
